changes the structure - moves classes to the suitable package - add more clarification to some class names
