+++ app/src/main/AndroidManifest.xml
@@ -17,10 +17,10 @@
 
         </activity>
         <activity
-            android:parentActivityName=".LoginActivity"
-            android:name=".SignUp" />
+            android:parentActivityName=".auth.LoginActivity"
+            android:name=".auth.SignUpActivity" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".auth.LoginActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
